Move event and update docs
